Refactor auth logic and reorganize components
Authentication API functions were moved from api.js to a new AuthService module for better separation of concerns. Configuration constants were extracted to apiConfig.js. Several components were reorganized into new folders, unused components (Filter, GetAPI, ResendOtpForm) were removed, and imports were updated accordingly throughout the codebase.
