+++ src/api/api.js
@@ -1,227 +1,20 @@
 import axios from "axios";
-
-const TOKEN = "8ed1bdc4-4501-11f0-9b81-222185cb68c8"; // Replace with valid token
-const GHN_BASE_URL = "https://dev-online-gateway.ghn.vn/shiip/public-api";
-export const SHOP_ID = 196808; // Replace with valid ShopID
-export const YOUR_SHOP_DISTRICT_ID = 1450; // Replace with valid district ID
-export const YOUR_SHOP_WARD_CODE = "20804"; // Replace with valid ward code
-
-// [0.] Các đường dẫn API
-// const API_BASE_URL = "https://localhost:7030/api";
-// const API_BASE_URL = "http://tom11357-001-site1.qtempurl.com/api";
-const API_BASE_URL = "https://tomsfurniturebackend.onrender.com/api";
-
-//#region [Global API🌐]-----------------------------------------------
-// [1.] Kiểm tra trạng thái đăng nhập của người dùng
-export const checkAuthStatus = async () => {
-    try {
-        // Gửi yêu cầu đến API để kiểm tra trạng thái đăng nhập + cookie
-        const response = await fetch(`${API_BASE_URL}/Auth/status`, {
-            method: "GET",
-            credentials: "include",
-        });
-        console.log("Auth status response:", response);
-        // Kiểm tra phản hồi từ server
-        if (!response.ok) {
-            const errorData = await response.json();
-            throw new Error(errorData.message || `Failed to check auth status: ${response.status}`);
-        }
-
-        const data = await response.json();
-        console.log("Auth status data:", data);
-        // Chuyển đổi dữ liệu từ API sang định dạng phù hợp
-        return {
-            isAuthenticated: data.isAuthenticated,
-            userName: data.userName || null,
-            email: data.email || null,
-            role: data.role || null,
-            message: data.message || null,
-            redirectUrl: data.redirectUrl || "/",
-        };
-    } catch (error) {
-        // Ghi log lỗi và trả về trạng thái mặc định
-        console.error("Lỗi khi kiểm tra trạng thái xác thực:", error);
-        return { isAuthenticated: false, message: "Unable to check authentication status." };
-    }
+import {
+    TOKEN,
+    GHN_BASE_URL,
+    SHOP_ID,
+    YOUR_SHOP_DISTRICT_ID,
+    YOUR_SHOP_WARD_CODE,
+    API_BASE_URL,
+} from "./apiConfig";
+export {
+    TOKEN,
+    GHN_BASE_URL,
+    SHOP_ID,
+    YOUR_SHOP_DISTRICT_ID,
+    YOUR_SHOP_WARD_CODE,
+    API_BASE_URL,
 };
-// [2.] Đăng nhập
-export const login = async (email, password) => {
-    try {
-        const response = await fetch(`${API_BASE_URL}/Auth/login`, {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            credentials: "include",
-            body: JSON.stringify({ email, password }),
-        });
-
-        const data = await response.json();
-        if (!response.ok) {
-            throw new Error(data.message || "Login failed.");
-        }
-
-        // Trả về dữ liệu từ API
-        return {
-            success: data.success || false,
-            message: data.message || "Login successful.",
-            userName: data.userName,
-            role: data.role,
-            redirectUrl: data.redirectUrl,
-        };
-    } catch (error) {
-        console.error("Lỗi khi đăng nhập:", error);
-        return { success: false, message: error.message || "An error occurred during login." };
-    }
-};
-// [3.] Đăng ký
-export const register = async (userData) => {
-    try {
-        const response = await fetch(`${API_BASE_URL}/Auth/register`, {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            credentials: "include",
-            body: JSON.stringify({
-                userName: userData.userName,
-                email: userData.email,
-                password: userData.password,
-                gender: userData.gender === "true", // Chuyển đổi string thành boolean
-                phoneNumber: userData.phoneNumber || null,
-                userAddress: userData.userAddress || null,
-            }),
-        });
-
-        const data = await response.json();
-        if (!response.ok) {
-            throw new Error(data.message || "Registration failed.");
-        }
-
-        return {
-            success: data.success || false,
-            message: data.message || "Registration successful.",
-        };
-    } catch (error) {
-        console.error("Lỗi khi đăng ký:", error);
-        return { success: false, message: error.message || "An error occurred during registration." };
-    }
-};
-// [4.] Xác thực OTP
-export const verifyOtp = async (email, otp) => {
-    try {
-        const response = await fetch(`${API_BASE_URL}/Auth/verify-otp`, {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            credentials: "include",
-            body: JSON.stringify({ email, otp }),
-        });
-
-        const data = await response.json();
-        if (!response.ok) {
-            throw new Error(data.message || "OTP verification failed.");
-        }
-
-        return {
-            success: data.success || false,
-            message: data.message || "OTP verification successful.",
-        };
-    } catch (error) {
-        console.error("Lỗi khi xác thực OTP:", error);
-        return { success: false, message: error.message || "An error occurred during OTP verification." };
-    }
-};
-// [5.] Gửi lại mã OTP
-export const resendOtp = async (email) => {
-    try {
-        const response = await fetch(`${API_BASE_URL}/Auth/resend-otp`, {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            credentials: "include",
-            body: JSON.stringify({ email }),
-        });
-
-        const data = await response.json();
-        if (!response.ok) {
-            throw new Error(data.message || "OTP resend failed.");
-        }
-
-        return {
-            success: data.success || false,
-            message: data.message || "New OTP sent successfully.",
-        };
-    } catch (error) {
-        console.error("Lỗi khi gửi lại OTP:", error);
-        return { success: false, message: error.message || "An error occurred during OTP resend." };
-    }
-};
-// [6.] Đăng xuất
-export const logout = async () => {
-    try {
-        const response = await fetch(`${API_BASE_URL}/Auth/logout`, {
-            method: "POST",
-            credentials: "include",
-        });
-
-        const data = await response.json();
-        if (!response.ok) {
-            throw new Error(data.message || "Logout failed.");
-        }
-
-        return {
-            success: data.success || false,
-            message: data.message || "Logout successful.",
-        };
-    } catch (error) {
-        console.error("Lỗi khi đăng xuất:", error);
-        return { success: false, message: error.message || "An error occurred during logout." };
-    }
-};
-// [7.] Yêu cầu quên mật khẩu
-export const forgotPassword = async (email) => {
-    try {
-        const response = await fetch(`${API_BASE_URL}/Auth/forgot-password`, {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            credentials: "include",
-            body: JSON.stringify({ email }),
-        });
-
-        const data = await response.json();
-        if (!response.ok) {
-            throw new Error(data.message || "Forgot password request failed.");
-        }
-
-        return {
-            success: data.success || false,
-            message: data.message || "OTP sent successfully.",
-        };
-    } catch (error) {
-        console.error("Lỗi khi yêu cầu quên mật khẩu:", error);
-        return { success: false, message: error.message || "An error occurred during forgot password request." };
-    }
-};
-// [8.] Đặt lại mật khẩu
-export const resetPassword = async (email, newPassword) => {
-    try {
-        const response = await fetch(`${API_BASE_URL}/Auth/reset-password`, {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            credentials: "include",
-            body: JSON.stringify({ email, newPassword }),
-        });
-
-        const data = await response.json();
-        if (!response.ok) {
-            throw new Error(data.message || "Password reset failed.");
-        }
-
-        return {
-            success: data.success || false,
-            message: data.message || "Password reset successfully.",
-        };
-    } catch (error) {
-        console.error("Lỗi khi đặt lại mật khẩu:", error);
-        return { success: false, message: error.message || "An error occurred during password reset." };
-    }
-};
-//#endregion [Global API🌐 - End]--------------------------------------
 
 //#region [Home Page 🏠]-----------------------------------------------
 export const getProductDetail = () => {
