+++ src/api/apiConfig.js
@@ -1,0 +1,8 @@
+export const TOKEN = "8ed1bdc4-4501-11f0-9b81-222185cb68c8"; // Replace with valid token
+export const GHN_BASE_URL = "https://dev-online-gateway.ghn.vn/shiip/public-api";
+export const SHOP_ID = 196808; // Replace with valid ShopID
+export const YOUR_SHOP_DISTRICT_ID = 1450; // Replace with valid district ID
+export const YOUR_SHOP_WARD_CODE = "20804"; // Replace with valid ward code
+// export const API_BASE_URL = "https://localhost:7030/api";
+// export const API_BASE_URL = "http://tom11357-001-site1.qtempurl.com/api";
+export const API_BASE_URL = "https://tomsfurniturebackend.onrender.com/api";
