--- src/components/Home/AuthComponents/ResendOtpForm.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export const ResendOtpForm = () => {
-    return <div>ResendOtpForm</div>;
-};
