--- src/components/Home/Filter.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Filter = () => {
-  return (
-    <div>Filter</div>
-  )
-}
-
-export default Filter
