+++ src/context/APIContext.jsx
@@ -64,7 +64,8 @@ export const APIProvider = ({ children }) => {
         if (!didInit) {
             fetchCategories();
             fetchStoreInformation();
-            fetchProducts({ pageNumber: 1, pageSize: 8 }); // Thêm pagination mặc định
+            // fetchProducts({ pageNumber: 1, pageSize: 8 }); // Thêm pagination mặc định
+            fetchProducts();
             fetchColors();
             fetchUnits();
             fetchMaterials();
