+++ src/routes/RouterSetup.jsx
@@ -38,8 +38,8 @@ import {
 import { createBrowserRouter, RouterProvider, Outlet, redirect, useNavigate, Navigate } from "react-router-dom";
 import HomeLayout from "@/pages/layouts/HomeLayout";
 import AdminLayouts from "@/pages/layouts/AdminLayouts";
+import AuthService from "@/api/service/AuthService";
 import {
-    checkAuthStatus,
     getAllBrands,
     getAllCategories,
     getAllColors,
@@ -90,7 +90,7 @@ const AdminRoute = ({ children }) => {
 
     useEffect(() => {
         const fetchAuthStatus = async () => {
-            const result = await checkAuthStatus();
+            const result = await AuthService.checkAuthStatus();
             setAuthStatus(result);
             if (!result.isAuthenticated || result.role !== "Admin") {
                 navigate("/");
